(fix) Make metrics aggregation independent of proxy logger
